# backend/api/static/swagger.yml
openapi: 3.0.0
info:
  title: API Token Management
  description: API for managing application tokens
  version: 1.0.0
servers:
  - url: http://localhost:5000/api/v1
paths:
...
  /applications:
    post:
      summary: Create a new application
      tags:
        - Applications
      security:
        - BearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationCreateRequest"
      responses:
        "201":
          description: Application created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Application created successfully"
                  app_id:
                    type: string
                    example: "123e4567-e89b-12d3-a456-426614174000"
        "400":
          description: Validation error
        "401":
          description: Unauthorized

    get:
      summary: Retrieve all applications
      tags:
        - Applications
      security:
        - BearerAuth: [ ]
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApplicationResponse"
        "401":
          description: Unauthorized
  /applications/{app_id}:
    put:
      summary: Update an application
      tags:
        - Applications
      security:
        - BearerAuth: [ ]
      parameters:
        - name: app_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdateRequest"
      responses:
        "200":
          description: Application updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Application updated successfully"
                  app_token:
                    type: string
                    example: "123e4567-e89b-12d3-a456-426614174000"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
        "404":
          description: Application not found
  /applications/log-connection:
    post:
      summary: Log a user's connection to an application
      tags:
        - Applications
      security:
        - BearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectionLogRequest"
      responses:
        "201":
          description: Connection log added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Connection log added successfully"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
  /applications/validate-token/{token}:
    get:
      summary: Validate an application token
      tags:
        - Applications
      security:
        - BearerAuth: [ ]
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  is_valid:
                    type: boolean
                    example: true
                  token:
                    type: string
                    example: "newly_generated_token_here"
        "401":
          description: Invalid token
  /api-tokens:
    post:
      summary: Generate a new API token
      tags:
        - API Tokens
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/APITokenRequest"
      responses:
        "201":
          description: Token created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APITokenResponse"
        "400":
          description: Validation error
        "401":
          description: Unauthorized

    get:
      summary: Retrieve all API tokens
      tags:
        - API Tokens
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of API tokens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/APITokenResponse"
        "401":
          description: Unauthorized
  /campaigns:
    post:
      summary: Create a new campaign
      tags:
        - Campaigns
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignCreateRequest"
      responses:
        "201":
          description: Campaign created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Campaign created successfully"
                  campaign_id:
                    type: integer
                    example: 1
        "400":
          description: Validation error
        "401":
          description: Unauthorized
  /campaigns/{app_id}:
    get:
      summary: Retrieve all campaigns linked to a specific application ID
      tags:
        - Campaigns
      security:
        - BearerAuth: []
      parameters:
        - name: app_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of campaigns for the given application ID
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CampaignResponse"
        "401":
          description: Unauthorized
  /campaigns/{campaign_id}:
    put:
      summary: Update a campaign name
      tags:
        - Campaigns
      security:
        - BearerAuth: []
      parameters:
        - name: campaign_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdateRequest"
      responses:
        "200":
          description: Campaign updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Campaign updated successfully"
                  campaign_id:
                    type: integer
                    example: 1
        "400":
          description: Validation error
        "401":
          description: Unauthorized
        "404":
          description: Campaign not found

    delete:
      summary: Delete a campaign by ID
      tags:
        - Campaigns
      security:
        - BearerAuth: []
      parameters:
        - name: campaign_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignDeleteRequest"
      responses:
        "200":
          description: Campaign deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Campaign deleted successfully"
                  campaign_id:
                    type: integer
                    example: 1
        "400":
          description: Validation error
        "401":
          description: Unauthorized
        "404":
          description: Campaign not found
  /api-tokens/{token}:
    delete:
      summary: Revoke an API token
      tags:
        - API Tokens
      security:
        - BearerAuth: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Token revoked successfully
        "401":
          description: Unauthorized
        "404":
          description: Token not found
  /api-tokens/rotate:
    put:
      summary: Rotate an API token
      tags:
        - API Tokens
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/APITokenRotateRequest"
      responses:
        "200":
          description: Token rotated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  new_token:
                    type: string
        "400":
          description: Validation error
        "401":
          description: Unauthorized
        "404":
          description: Token not found
components:
  schemas:
...
    ApplicationCreateRequest:
      type: object
      properties:
        name:
          type: string
          example: "MyApplication"
        mail_sender:
          type: string
          example: "no-reply@myapp.com"
      required:
        - name
    ApplicationResponse:
      type: object
      properties:
        id:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          example: "MyApplication"
        created_by:
          type: string
          example: "user123"
        mail_sender:
          type: string
          example: "no-reply@myapp.com"
        created_at:
          type: string
          format: date-time
          example: "2025-01-01T12:00:00Z"
    ApplicationUpdateRequest:
      type: object
      properties:
        name:
          type: string
          example: "UpdatedAppName"
        new_mail_sender:
          type: string
          example: "updated-email@myapp.com"
        generate_new_id:
          type: boolean
          example: true
    ConnectionLogRequest:
      type: object
      properties:
        app_ids:
          type: array
          items:
            type: string
          example: ["app-name", "another-app"]
      required:
        - app_ids
    CampaignCreateRequest:
      type: object
      properties:
        name:
          type: string
          example: "Marketing Campaign"
        app_id:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
      required:
        - name
        - app_id
    CampaignResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Marketing Campaign"
        app_id:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
        created_by:
          type: string
          example: "user123"
        created_at:
          type: string
          format: date-time
          example: "2025-01-01T12:00:00Z"
    CampaignUpdateRequest:
      type: object
      properties:
        name:
          type: string
          example: "Updated Campaign Name"
        app_id:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
    CampaignDeleteRequest:
      type: object
      properties:
        app_id:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
      required:
        - app_id
    APITokenRequest:
      type: object
      properties:
        token_name:
          type: string
          example: "MyAppToken"
        app_names:
          type: array
          items:
            type: string
          example: ["App1", "App2"]
        expiration:
          type: string
          format: date-time
          example: "2025-12-31T23:59:59Z"
      required:
        - token_name
        - app_names
        - expiration
    APITokenResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        token_name:
          type: string
          example: "MyAppToken"
        app_names:
          type: array
          items:
            type: string
          example: ["App1", "App2"]
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR..."
        created_by:
          type: string
          example: "user123"
        expiration:
          type: string
          format: date-time
          example: "2025-12-31T23:59:59Z"
        created_at:
          type: string
          format: date-time
          example: "2025-01-01T12:00:00Z"
    APITokenRotateRequest:
      type: object
      properties:
        token_name:
          type: string
          example: "MyAppToken"
      required:
        - token_name
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT